add  DOM to delete + modify items from the list to projects
