+++ src/index.js
@@ -1,8 +1,10 @@
 import './styles.scss';
 
 import { projectFactory } from './factory';
+import { initialize, saveData } from './storagedata';
+import { displayProjects } from './display';
+// import { modifyItem } from './modify';
 
-import { initialize, saveData} from './storagedata';
 
 
 const initialData = initialize([], 0);
@@ -39,51 +41,51 @@ const todoFactory = (title, duedate, desc, note, priority, temp = 'Empty') => {
 
 
 ///  dispaly the To-do list's
-const displayProject = () => {
-  const projectsMain = document.querySelector('#all-projects-content');
-  document.querySelector('#all-projects-content').innerHTML = '';
+// const displayProjects = (projects) => {
+//   const projectsMain = document.querySelector('#all-projects-content');
+//   document.querySelector('#all-projects-content').innerHTML = '';
 
-  projects.forEach((project) => {
-    const container = document.createElement('div');
-    container.setAttribute('class', `container ${project.name}-project`);
+//   projects.forEach((project) => {
+//     const container = document.createElement('div');
+//     container.setAttribute('class', `container ${project.name}-project`);
 
-    const heading = document.createElement('h3');
-    heading.textContent = project.name;
+//     const heading = document.createElement('h3');
+//     heading.textContent = project.name;
 
-    const listElement = document.createElement('ul');
-    listElement.id = `${project.name}-list`;
+//     const listElement = document.createElement('ul');
+//     listElement.id = `${project.name}-list`;
 
-    container.append(heading, listElement);
-    projectsMain.appendChild(container);
+//     container.append(heading, listElement);
+//     projectsMain.appendChild(container);
 
-    const listItems = project.list;
-    listItems.forEach(item => {
-      const listItem = document.createElement('li');
-      const customDate = formatDate(item.duedate);
-      // Saturday, September 17, 2016
-      listItem.textContent = `${item.title}, Date: ${customDate}, Priority: ${item.priority}`;
-      // const testBtn = document.createElement('button');
-      const modifyBtn = document.createElement('button');
-      const deleteBtn = document.createElement('button');
-      modifyBtn.innerHTML = '<img src="https://img.icons8.com/fluent-systems-regular/15/000000/edit-property.png" />';
-      deleteBtn.innerHTML = '<img src="https://img.icons8.com/material-sharp/15/000000/delete-forever.png" />';
+//     const listItems = project.list;
+//     listItems.forEach(item => {
+//       const listItem = document.createElement('li');
+//       const customDate = formatDate(item.duedate);
+//       // Saturday, September 17, 2016
+//       listItem.textContent = `${item.title}, Date: ${customDate}, Priority: ${item.priority}`;
+//       // const testBtn = document.createElement('button');
+//       const modifyBtn = document.createElement('button');
+//       const deleteBtn = document.createElement('button');
+//       modifyBtn.innerHTML = '<img src="https://img.icons8.com/fluent-systems-regular/15/000000/edit-property.png" />';
+//       deleteBtn.innerHTML = '<img src="https://img.icons8.com/material-sharp/15/000000/delete-forever.png" />';
 
-      // const expand = document.querySelector('modifyBtn');
-      // const loose = document.querySelector('deleteBtn');
+//       // const expand = document.querySelector('modifyBtn');
+//       // const loose = document.querySelector('deleteBtn');
 
-      // if (modifyBtn != null ){
-      //   modifyBtn.addEventListener('onclick', modifyItem);
-      // }
-      modifyBtn.onclick = () => modifyItem(item, project);
-      deleteBtn.onclick = () => deleteItem(item, project);
+//       // if (modifyBtn != null ){
+//       //   modifyBtn.addEventListener('onclick', modifyItem);
+//       // }
+//       modifyBtn.onclick = () => modifyItem(item, project);
+//       deleteBtn.onclick = () => deleteItem(item, project);
 
-      // loose.addEventListener('onclick', deleteItem);
+//       // loose.addEventListener('onclick', deleteItem);
 
-      listItem.append(modifyBtn, deleteBtn);
-      listElement.appendChild(listItem);
-    });
-  });
-};
+//       listItem.append(modifyBtn, deleteBtn);
+//       listElement.appendChild(listItem);
+//     });
+//   });
+// };
 
 // delete task from projects
 
@@ -96,7 +98,7 @@ const deleteItem = (task, project) => {
   // projects = projects.filter(i => i != currentProject);
   // projects.push(newProject);
   saveData(projects, id);
-  displayProject();
+  displayProjects(projects);
 };
 
 // const deleteOption = document.querySelector('deleteBtn');
@@ -153,7 +155,7 @@ const saveModifiedData = (item, project) => {
     deleteItem(oldTask, oldProject);
   }
   saveData(projects, id);
-  displayProject();
+  displayProjects(projects);
 
   return false;
 };
@@ -193,15 +195,15 @@ const modifyItem = (item, project) => {
 };
 
 // create the input due date
-const formatDate = (input) => {
-  const options = {
-    year: 'numeric', month: 'long', day: 'numeric',
-  };
+// const formatDate = (input) => {
+//   const options = {
+//     year: 'numeric', month: 'long', day: 'numeric',
+//   };
 
-  const date = new Date(input);
-  const result = date.toLocaleDateString('en-US', options);
-  return result;
-};
+//   const date = new Date(input);
+//   const result = date.toLocaleDateString('en-US', options);
+//   return result;
+// };
 
 if (projects.length === 0) {
   const defaultProject = projectFactory('default');
@@ -211,7 +213,7 @@ if (projects.length === 0) {
   console.log('initialize');
   console.log({ projects });
 } else {
-  displayProject();
+  displayProjects(projects);
 }
 
 // const room1 = todoFactory('Clean my room', '1/2/2021', 'test', 'a', 'Low');
@@ -227,7 +229,7 @@ const addTaskToProject = (task, project) => {
   const currentProject = projects.find(o => o.name === project);
   currentProject.list.push(task);
   saveData(projects, id);
-  displayProject(project);
+  displayProjects(projects);
 };
 
 //  store the list of tasks
@@ -277,3 +279,5 @@ const btn = document.querySelector('#tasksubmit');
 
 btn.onclick = forminput;
 // document.querySelector('button').addEventListener('onclick', (event) => forminput);
+
+export { modifyItem, deleteItem};
